+++ spec/rigor/configuration_spec.rb
@@ -12,6 +12,38 @@ RSpec.describe Rigor::Configuration do
         expect(configuration.paths).to eq(["lib"])
         expect(configuration.plugins).to eq([])
         expect(configuration.cache_path).to eq(".rigor/cache")
+        expect(configuration.libraries).to eq([])
+        expect(configuration.signature_paths).to be_nil
+      end
+    end
+
+    it "reads libraries: and signature_paths: from the YAML file" do
+      Dir.mktmpdir do |dir|
+        path = File.join(dir, ".rigor.yml")
+        File.write(path, <<~YAML)
+          libraries:
+            - csv
+            - set
+          signature_paths:
+            - sig
+            - vendor/sig
+        YAML
+
+        configuration = described_class.load(path)
+
+        expect(configuration.libraries).to eq(%w[csv set])
+        expect(configuration.signature_paths).to eq(%w[sig vendor/sig])
+      end
+    end
+
+    it "treats signature_paths: [] as 'load no project signatures'" do
+      Dir.mktmpdir do |dir|
+        path = File.join(dir, ".rigor.yml")
+        File.write(path, "signature_paths: []\n")
+
+        configuration = described_class.load(path)
+
+        expect(configuration.signature_paths).to eq([])
       end
     end
   end